requirements/buildbot.txt
mercurial==4.4.1
Django==1.11.7
#Twisted==8.2.0
./zope.interface-4.4.3-cp27-cp27mu-manylinux1_x86_64.whl
./Twisted-8.2.0-cp27-cp27mu-linux_x86_64.whl
./buildbot-0.7.12-py2-none-any.whl
simplejson
MySQL-python
elasticsearch==1.2
compare-locales==2.7.0
fluent==0.6.2
markus[datadog]==1.1
